Refactor: wallet testing tabs
