+++ src/components/WalletTest.svelte
@@ -23,6 +23,9 @@
   import ConnectorModal from './Modals/ConnectorModal.svelte';
   import ChainSwitchModal from './Modals/ChainSwitchModal.svelte';
   import EIPResultsModal from './Modals/EIPResultsModal.svelte';
+  import TransactionsTab from './Tabs/TransactionsTab.svelte';
+  import SignaturesTab from './Tabs/SignaturesTab.svelte';
+  import EIPSupportTab from './Tabs/EIPSupportTab.svelte';
 
   type Account = ReturnType<typeof getAccount>;
 
@@ -768,366 +771,28 @@ Issued At: ${new Date().toISOString()}`;
 
     <!-- Main Content -->
     <div class="main-content">
-      {#if uiState.activeTab === 'transactions' && uiState.selectedTxId}
+      {#if uiState.activeTab === 'transactions'}
         {@const selectedTx = testTransactions.find((tx) => tx.id === uiState.selectedTxId)}
-        {#if selectedTx}
-          {@const isActive = transactionState.activeId === selectedTx.id}
-          {@const isPending = transactionState.isPending && isActive}
-          {@const txHash = transactionState.hashes[selectedTx.id]}
-          <div class="detail-card" data-card="radius-8 padding-5">
-            <header data-row="gap-2 start wrap">
-              <div data-column="gap-1">
-                <h3>{selectedTx.name}</h3>
-                {#if selectedTx.description}
-                  <p class="body-text">{selectedTx.description}</p>
-                {/if}
-              </div>
-              {#if txHash}
-                <button
-                  type="button"
-                  class="explorer-link"
-                  onclick={() => openInExplorer(txHash)}
-                  title="View on Etherscan"
-                >
-                  ↗
-                </button>
-              {/if}
-            </header>
-
-            <div data-column="gap-4">
-              {#if selectedTx.requirements && selectedTx.requirements.length > 0}
-                <div class="requirements-box">
-                  <h4 class="requirements-title">📋 Requirements:</h4>
-                  <ul class="requirements-list">
-                    {#each selectedTx.requirements as requirement}
-                      <li>{requirement}</li>
-                    {/each}
-                  </ul>
-                </div>
-              {/if}
-
-              <div class="detail-section">
-                <span class="detail-label">📞 Function:</span>
-                <code class="detail-code">{selectedTx.function}</code>
-              </div>
-
-              {#if selectedTx.parameters.length > 0}
-                <div class="detail-section">
-                  <span class="detail-label">📋 Parameters:</span>
-                  <div class="parameters-list" data-column="gap-2">
-                    {#each selectedTx.parameters as param}
-                      <div class="parameter-item">
-                        <div class="parameter-header">
-                          <span class="parameter-name">{param.name}:</span>
-                          <span class="parameter-type">({param.type})</span>
-                        </div>
-                        <code class="parameter-value">{formatValue(param.value, param.type)}</code>
-                      </div>
-                    {/each}
-                  </div>
-                </div>
-              {/if}
-
-              {#if selectedTx.calls && selectedTx.calls.length > 0}
-                <div class="detail-section">
-                  <span class="detail-label">🔗 Calls ({selectedTx.calls.length}):</span>
-                  <div class="parameters-list" data-column="gap-3">
-                    {#each selectedTx.calls as call, idx}
-                      <div class="parameter-item">
-                        <div class="parameter-header">
-                          <span class="parameter-name">Call {idx + 1}</span>
-                        </div>
-                        <div data-column="gap-2">
-                          <div>
-                            <span class="parameter-type">To:</span>
-                            <code class="parameter-value">{call.to}</code>
-                          </div>
-                          <div>
-                            <span class="parameter-type">Data:</span>
-                            <code class="parameter-value calldata">{call.data}</code>
-                          </div>
-                          {#if call.value !== undefined && call.value > 0n}
-                            <div>
-                              <span class="parameter-type">Value:</span>
-                              <code class="parameter-value">{call.value.toString()} Wei</code>
-                            </div>
-                          {/if}
-                        </div>
-                      </div>
-                    {/each}
-                  </div>
-                </div>
-              {:else}
-                <div class="detail-section">
-                  <span class="detail-label">📦 Calldata:</span>
-                  <code class="detail-code calldata">{selectedTx.calldata}</code>
-                </div>
-
-                {#if selectedTx.contractAddress}
-                  <div class="detail-section">
-                    <span class="detail-label">📍 Contract Address:</span>
-                    <code class="detail-code">{selectedTx.contractAddress}</code>
-                  </div>
-                {/if}
-              {/if}
-
-              <div class="warning-box">
-                <p class="warning-text">
-                  <strong>⚠️ WARNING:</strong> This page is for testing only. Do NOT send real transactions.
-                </p>
-              </div>
-
-              <button
-                type="button"
-                data-pressable
-                onclick={() => handleSendTransaction(selectedTx)}
-                disabled={!account?.address || isPending || (!selectedTx.calls && !selectedTx.contractAddress) || (selectedTx.calls && selectedTx.calls.length === 0)}
-              >
-                {#if isPending}
-                  Preparing…
-                {:else if txHash}
-                  Transaction Sent ✓
-                {:else}
-                  Send Transaction (Testing Only)
-                {/if}
-              </button>
-
-              {#if txHash}
-                <div class="result-box">
-                  <span class="result-label">Transaction Hash:</span>
-                  <button type="button" class="result-link" onclick={() => openInExplorer(txHash)}>
-                    {txHash.slice(0, 10)}…{txHash.slice(-8)} ↗
-                  </button>
-                </div>
-              {/if}
-            </div>
-          </div>
-        {/if}
-      {:else if uiState.activeTab === 'signatures' && uiState.selectedSigId}
+        <TransactionsTab
+          {selectedTx}
+          {transactionState}
+          {account}
+          onSendTransaction={handleSendTransaction}
+          onOpenInExplorer={openInExplorer}
+          {formatValue}
+        />
+      {:else if uiState.activeTab === 'signatures'}
         {@const selectedSig = testSignatures.find((sig) => sig.id === uiState.selectedSigId)}
-        {#if selectedSig}
-          {@const isActive = signatureState.activeId === selectedSig.id}
-          {@const isPending = signatureState.isPending && isActive}
-          {@const sigResult = signatureState.results[selectedSig.id]}
-          <div class="detail-card" data-card="radius-8 padding-5">
-            <header data-row="gap-2 start wrap">
-              <div data-column="gap-1">
-                <h3>{selectedSig.name}</h3>
-                {#if selectedSig.description}
-                  <p class="body-text">{selectedSig.description}</p>
-                {/if}
-              </div>
-            </header>
-
-            <div data-column="gap-4">
-              {#if selectedSig.requirements && selectedSig.requirements.length > 0}
-                <div class="requirements-box">
-                  <h4 class="requirements-title">📋 Safety Notes:</h4>
-                  <ul class="requirements-list">
-                    {#each selectedSig.requirements as requirement}
-                      <li>{requirement}</li>
-                    {/each}
-                  </ul>
-                </div>
-              {/if}
-
-              <div class="detail-section">
-                <span class="detail-label">📝 Signature Type:</span>
-                <code class="detail-code">
-                  {selectedSig.type === 'message' ? 'Simple Message' : 'EIP-712 Typed Data'}
-                </code>
-              </div>
-
-              {#if selectedSig.type === 'message' && selectedSig.message}
-                <div class="detail-section">
-                  <span class="detail-label">💬 Message:</span>
-                  <code class="detail-code message-content">{selectedSig.message}</code>
-                </div>
-              {/if}
-
-              {#if selectedSig.type === 'typed' && selectedSig.domain}
-                <div class="detail-section">
-                  <span class="detail-label">🏷️ Domain:</span>
-                  <div class="domain-box">
-                    <div class="domain-item">
-                      <span class="domain-key">Name:</span> {selectedSig.domain.name}
-                    </div>
-                    <div class="domain-item">
-                      <span class="domain-key">Version:</span> {selectedSig.domain.version}
-                    </div>
-                    <div class="domain-item">
-                      <span class="domain-key">Chain ID:</span> {selectedSig.domain.chainId}
-                    </div>
-                    <div class="domain-item">
-                      <span class="domain-key">Verifying Contract:</span> {selectedSig.domain.verifyingContract}
-                    </div>
-                    <div class="domain-item">
-                      <span class="domain-key">Salt:</span> {selectedSig.domain.salt}
-                    </div>
-                  </div>
-                </div>
-
-                {#if selectedSig.primaryType}
-                  <div class="detail-section">
-                    <span class="detail-label">📋 Primary Type:</span>
-                    <code class="detail-code">{selectedSig.primaryType}</code>
-                  </div>
-                {/if}
-
-                {#if selectedSig.messageData}
-                  <div class="detail-section">
-                    <span class="detail-label">💬 Message Data:</span>
-                    <code class="detail-code message-content">
-                      {JSON.stringify(selectedSig.messageData, null, 2)}
-                    </code>
-                  </div>
-                {/if}
-              {/if}
-
-              <div class="warning-box warning-yellow">
-                <p class="warning-text">
-                  <strong>⚠️ WARNING:</strong> Only sign messages you trust. This page is for testing and
-                  educational purposes only.
-                </p>
-              </div>
-
-              <button
-                type="button"
-                data-pressable
-                onclick={() => {
-                  if (selectedSig.type === 'message') {
-                    void handleSignMessage(selectedSig);
-                  } else {
-                    void handleSignTypedData(selectedSig);
-                  }
-                }}
-                disabled={!account?.address || isPending}
-              >
-                {#if isPending}
-                  Signing…
-                {:else if sigResult}
-                  Signature Created ✓
-                {:else}
-                  Sign {selectedSig.type === 'message' ? 'Message' : 'Typed Data'}
-                {/if}
-              </button>
-
-              {#if sigResult}
-                <div class="result-box">
-                  <span class="result-label">Signature:</span>
-                  <code class="signature-result">{sigResult}</code>
-                </div>
-              {/if}
-            </div>
-          </div>
-        {/if}
-      {:else if uiState.activeTab === 'eip-support' && uiState.selectedEipId}
+        <SignaturesTab
+          {selectedSig}
+          {signatureState}
+          {account}
+          onSignMessage={handleSignMessage}
+          onSignTypedData={handleSignTypedData}
+        />
+      {:else if uiState.activeTab === 'eip-support'}
         {@const selectedEip = eipTests.find((eip) => eip.id === uiState.selectedEipId)}
-        {#if selectedEip}
-          {@const isActive = eipState.activeId === selectedEip.id}
-          {@const isTesting = eipState.isTesting && isActive}
-          {@const testResults = eipState.results[selectedEip.id]}
-          <div class="detail-card" data-card="radius-8 padding-5">
-            <header data-row="gap-2 start wrap">
-              <div data-column="gap-1">
-                <h3>{selectedEip.eipNumber}: {selectedEip.name}</h3>
-                {#if selectedEip.description}
-                  <p class="body-text">{selectedEip.description}</p>
-                {/if}
-              </div>
-              <a
-                href={selectedEip.specUrl}
-                target="_blank"
-                rel="noopener noreferrer"
-                class="spec-link"
-                title="View specification"
-              >
-                Spec ↗
-              </a>
-            </header>
-
-            <div data-column="gap-4">
-              {#if selectedEip.requirements && selectedEip.requirements.length > 0}
-                <div class="requirements-box">
-                  <h4 class="requirements-title">📋 Requirements:</h4>
-                  <ul class="requirements-list">
-                    {#each selectedEip.requirements as requirement}
-                      <li>{requirement}</li>
-                    {/each}
-                  </ul>
-                </div>
-              {/if}
-
-              <div class="detail-section">
-                <span class="detail-label">🧪 Compliance Checks:</span>
-                <div class="eip-checks" data-column="gap-2">
-                  {#each selectedEip.checks as check}
-                    {@const status = testResults?.[check.id] || 'untested'}
-                    <div class="check-item" data-row="gap-2 start">
-                      <span class="check-status status-{status}" title={status}>
-                        {#if status === 'pass'}
-                          ✓
-                        {:else if status === 'fail'}
-                          ✗
-                        {:else if status === 'partial'}
-                          ◐
-                        {:else}
-                          ○
-                        {/if}
-                      </span>
-                      <div data-column="gap-1">
-                        <div class="check-name">
-                          {check.name}
-                          {#if check.critical}
-                            <span class="critical-badge">required</span>
-                          {/if}
-                        </div>
-                        <div class="check-description">{check.description}</div>
-                      </div>
-                    </div>
-                  {/each}
-                </div>
-              </div>
-
-              {#if eipState.discoveredProviders.length > 0 && selectedEip.id === 'eip-6963'}
-                <div class="detail-section">
-                  <span class="detail-label">🔍 Discovered Providers ({eipState.discoveredProviders.length}):</span>
-                  <div class="providers-list" data-column="gap-2">
-                    {#each eipState.discoveredProviders as provider}
-                      <div class="provider-item">
-                        <div class="provider-header" data-row="gap-2 start">
-                          {#if provider.icon}
-                            <img src={provider.icon} alt={provider.name} class="provider-icon" />
-                          {/if}
-                          <div>
-                            <div class="provider-name">{provider.name}</div>
-                            <div class="provider-rdns">{provider.rdns}</div>
-                          </div>
-                        </div>
-                        <div class="provider-uuid">UUID: {provider.uuid}</div>
-                      </div>
-                    {/each}
-                  </div>
-                </div>
-              {/if}
-
-              <button
-                type="button"
-                data-pressable
-                onclick={() => testEIPSupport(selectedEip)}
-                disabled={isTesting}
-              >
-                {#if isTesting}
-                  Testing…
-                {:else if testResults}
-                  Re-run Tests
-                {:else}
-                  Run Tests
-                {/if}
-              </button>
-            </div>
-          </div>
-        {/if}
+        <EIPSupportTab {selectedEip} {eipState} onTestEIPSupport={testEIPSupport} />
       {/if}
     </div>
   </div>
@@ -1237,211 +902,6 @@ Issued At: ${new Date().toISOString()}`;
     min-width: 0;
   }
 
-  .detail-card h3 {
-    font-size: 1.1rem;
-    font-weight: 600;
-    margin: 0 0 0.5rem 1rem;
-  }
-
-  .body-text {
-    font-size: 0.9rem;
-    color: var(--text-secondary);
-    margin: 0;
-  }
-
-  .explorer-link {
-    background: transparent;
-    border: none;
-    color: var(--accent);
-    font-size: 1.2rem;
-    cursor: pointer;
-    padding: 0.25rem;
-    margin-left: auto;
-  }
-
-  .explorer-link:hover {
-    opacity: 0.8;
-  }
-
-  .requirements-box {
-    padding: 1rem;
-    background: color-mix(in srgb, var(--accent) 10%, transparent);
-    border: 1px solid color-mix(in srgb, var(--accent) 30%, transparent);
-    border-radius: 0.5rem;
-  }
-
-  .requirements-title {
-    font-size: 0.85rem;
-    font-weight: 600;
-    color: var(--accent);
-    margin: 0 0 0.75rem 0;
-  }
-
-  .requirements-list {
-    margin: 0;
-    padding-left: 1.25rem;
-    display: flex;
-    flex-direction: column;
-    gap: 0.5rem;
-  }
-
-  .requirements-list li {
-    font-size: 0.85rem;
-    color: var(--text-secondary);
-  }
-
-  .detail-section {
-    display: flex;
-    flex-direction: column;
-    gap: 0.5rem;
-  }
-
-  .detail-label {
-    font-size: 0.8rem;
-    font-weight: 500;
-    color: var(--text-secondary);
-  }
-
-  .detail-code {
-    font-family: ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, 'Liberation Mono', 'Courier New',
-      monospace;
-    font-size: 0.8rem;
-    background-color: var(--background-secondary);
-    padding: 0.75rem;
-    border-radius: 0.5rem;
-    word-break: break-all;
-    color: var(--text-primary);
-  }
-
-  .detail-code.calldata {
-    font-size: 0.75rem;
-  }
-
-  .detail-code.message-content {
-    white-space: pre-wrap;
-    font-size: 0.75rem;
-  }
-
-  .parameters-list {
-    margin-top: 0.5rem;
-  }
-
-  .parameter-item {
-    padding: 0.75rem;
-    background: color-mix(in srgb, var(--background-secondary) 50%, transparent);
-    border: 1px solid var(--background-secondary);
-    border-radius: 0.5rem;
-  }
-
-  .parameter-header {
-    display: flex;
-    gap: 0.5rem;
-    margin-bottom: 0.5rem;
-  }
-
-  .parameter-name {
-    font-size: 0.8rem;
-    font-weight: 500;
-    color: var(--accent);
-  }
-
-  .parameter-type {
-    font-size: 0.75rem;
-    color: var(--text-secondary);
-  }
-
-  .parameter-value {
-    font-family: ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, 'Liberation Mono', 'Courier New',
-      monospace;
-    font-size: 0.75rem;
-    color: var(--text-primary);
-    word-break: break-all;
-  }
-
-  .domain-box {
-    padding: 0.75rem;
-    background: var(--background-secondary);
-    border-radius: 0.5rem;
-    display: flex;
-    flex-direction: column;
-    gap: 0.5rem;
-  }
-
-  .domain-item {
-    font-size: 0.8rem;
-    color: var(--text-primary);
-  }
-
-  .domain-key {
-    font-weight: 500;
-    color: var(--accent);
-  }
-
-  .warning-box {
-    padding: 0.75rem;
-    background: color-mix(in srgb, var(--rating-fail) 10%, transparent);
-    border: 1px solid color-mix(in srgb, var(--rating-fail) 30%, transparent);
-    border-radius: 0.5rem;
-  }
-
-  .warning-box.warning-yellow {
-    background: color-mix(in srgb, #fbbf24 10%, transparent);
-    border-color: color-mix(in srgb, #fbbf24 30%, transparent);
-  }
-
-  .warning-text {
-    font-size: 0.8rem;
-    color: var(--rating-fail);
-    margin: 0;
-    text-align: center;
-  }
-
-  .warning-box.warning-yellow .warning-text {
-    color: #fbbf24;
-  }
-
-  .result-box {
-    padding: 0.75rem;
-    background: var(--background-secondary);
-    border: 1px solid color-mix(in srgb, var(--accent) 30%, transparent);
-    border-radius: 0.5rem;
-    display: flex;
-    flex-direction: column;
-    gap: 0.5rem;
-  }
-
-  .result-label {
-    font-size: 0.8rem;
-    font-weight: 500;
-    color: var(--text-secondary);
-  }
-
-  .result-link {
-    background: transparent;
-    border: none;
-    color: var(--accent);
-    font-size: 0.8rem;
-    cursor: pointer;
-    text-align: left;
-    font-family: ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, 'Liberation Mono', 'Courier New',
-      monospace;
-  }
-
-  .result-link:hover {
-    opacity: 0.8;
-  }
-
-  .signature-result {
-    font-family: ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, 'Liberation Mono', 'Courier New',
-      monospace;
-    font-size: 0.75rem;
-    background: color-mix(in srgb, var(--background-primary) 50%, transparent);
-    padding: 0.75rem;
-    border-radius: 0.5rem;
-    word-break: break-all;
-    color: var(--text-primary);
-  }
-
   .account-pill {
     display: inline-flex;
     align-items: center;
@@ -1462,124 +922,5 @@ Issued At: ${new Date().toISOString()}`;
 
   button[data-pressable] {
     min-width: 0;
-  }
-
-  .spec-link {
-    background: transparent;
-    border: none;
-    color: var(--accent);
-    font-size: 0.9rem;
-    cursor: pointer;
-    text-decoration: none;
-    margin-left: auto;
-  }
-
-  .spec-link:hover {
-    opacity: 0.8;
-  }
-
-  .eip-checks {
-    margin-top: 0.5rem;
-  }
-
-  .check-item {
-    padding: 0.75rem;
-    background: color-mix(in srgb, var(--background-secondary) 50%, transparent);
-    border: 1px solid var(--background-secondary);
-    border-radius: 0.5rem;
-  }
-
-  .check-status {
-    font-size: 1.2rem;
-    font-weight: bold;
-    width: 1.5rem;
-    height: 1.5rem;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border-radius: 50%;
-    flex-shrink: 0;
-  }
-
-  .check-status.status-pass {
-    color: var(--rating-pass);
-    background: color-mix(in srgb, var(--rating-pass) 20%, transparent);
-  }
-
-  .check-status.status-fail {
-    color: var(--rating-fail);
-    background: color-mix(in srgb, var(--rating-fail) 20%, transparent);
-  }
-
-  .check-status.status-partial {
-    color: var(--rating-partial);
-    background: color-mix(in srgb, var(--rating-partial) 20%, transparent);
-  }
-
-  .check-status.status-untested {
-    color: var(--text-secondary);
-    background: color-mix(in srgb, var(--background-secondary) 50%, transparent);
-  }
-
-  .check-name {
-    font-size: 0.85rem;
-    font-weight: 500;
-    color: var(--text-primary);
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-  }
-
-  .check-description {
-    font-size: 0.75rem;
-    color: var(--text-secondary);
-  }
-
-  .critical-badge {
-    font-size: 0.65rem;
-    padding: 0.125rem 0.375rem;
-    background: color-mix(in srgb, var(--rating-fail) 20%, transparent);
-    color: var(--rating-fail);
-    border-radius: 0.25rem;
-    font-weight: 600;
-    text-transform: uppercase;
-  }
-
-  .providers-list {
-    margin-top: 0.5rem;
-  }
-
-  .provider-item {
-    padding: 0.75rem;
-    background: color-mix(in srgb, var(--background-secondary) 50%, transparent);
-    border: 1px solid var(--background-secondary);
-    border-radius: 0.5rem;
-  }
-
-  .provider-icon {
-    width: 2rem;
-    height: 2rem;
-    border-radius: 0.25rem;
-  }
-
-  .provider-name {
-    font-size: 0.85rem;
-    font-weight: 500;
-    color: var(--text-primary);
-  }
-
-  .provider-rdns {
-    font-size: 0.75rem;
-    color: var(--text-secondary);
-    font-family: ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, 'Liberation Mono', 'Courier New',
-      monospace;
-  }
-
-  .provider-uuid {
-    font-size: 0.7rem;
-    color: var(--text-secondary);
-    margin-top: 0.5rem;
-    font-family: ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, 'Liberation Mono', 'Courier New',
-      monospace;
   }
 </style>
